requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alabaster==0.7.12
antlr4-python3-runtime==4.9.2
argparse==1.4.0
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
babel==2.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
bcp47==0.0.4
beautifulsoup4==4.9.3
bioregistry==0.2.6
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cfgraph==0.2.1
charset-normalizer==2.0.1; python_version >= '3'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
commonmark==0.9.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
deprecated==1.2.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
docopt==0.6.2
docutils==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
frozendict==2.0.3; python_version >= '3.6'
funowl==0.1.6
ghp-import==2.0.1
googlemaps==4.4.5
graphviz==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
greenlet==1.1.0; python_version >= '3'
hbreader==0.9.1; python_version >= '3.7'
idna==3.2; python_version >= '3'
imagesize==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.6.1; python_version >= '3.6'
inflect==5.3.0; python_version >= '3.6'
isodate==0.6.0
jinja2==3.0.1; python_version >= '3.6'
jsonasobj2==1.0.4; python_version >= '3.6'
jsonasobj==2.0.2.dev4; python_version >= '3.6'
jsonschema==4.0.0a2
linkml-model==1.0.1
linkml-runtime==1.0.10
linkml==1.0.1
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown==3.3.4; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
mergedeep==1.3.4; python_version >= '3.6'
mkdocs==1.2.1; python_version >= '3.6'
more-click==0.0.5; python_version >= '3.6'
more-itertools==8.8.0; python_version >= '3.5'
nmdc-schema==2021.7.1rc1
num2words==0.5.10
numpy==1.21.0; python_version >= '3.7'
packaging==21.0; python_version >= '3.6'
pandas==1.3.0
parse==1.19.0
pint==0.17
prefixcommons==0.1.9
prologterms==0.0.6
pydantic==1.8.2; python_full_version >= '3.6.1'
pygments==2.9.0; python_version >= '3.5'
pyjsg==0.11.6
pyldmod==2.0.5
pyparsing==3.0.0b2; python_version >= '3.5'
pyrsistent==0.18.0; python_version >= '3.6'
pyshex==0.7.20; python_version >= '3.6'
pyshexc==0.8.3; python_version >= '3.7'
pystow==0.1.7; python_version >= '3.6'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml-env-tag==0.1; python_version >= '3.6'
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
quantulum3==0.7.7
rdflib-jsonld==0.5.0
rdflib-pyldmod-compat==0.1.2
rdflib==5.0.0
recommonmark==0.7.1
requests-ftp==0.3.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rfc3987==1.3.8
shexjsg==0.7.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
snowballstemmer==2.1.0
soupsieve==2.2.1; python_version >= '3'
sparqlslurper==0.4.1; python_version >= '3.6'
sparqlwrapper==1.8.5
sphinx-rtd-theme==0.5.2
sphinx==4.1.0
sphinxcontrib-applehelp==1.0.2; python_version >= '3.5'
sphinxcontrib-devhelp==1.0.2; python_version >= '3.5'
sphinxcontrib-htmlhelp==2.0.0; python_version >= '3.6'
sphinxcontrib-jsmath==1.0.1; python_version >= '3.5'
sphinxcontrib-qthelp==1.0.3; python_version >= '3.5'
sphinxcontrib-serializinghtml==1.1.5; python_version >= '3.5'
sqlalchemy==1.4.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
tabulate==0.8.9
testfixtures==6.17.1
tqdm==4.61.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
watchdog==2.1.3; python_version >= '3.6'
wrapt==1.12.1
zipp==3.5.0; python_version >= '3.6'
runner@git+https://github.com/monarch-initiative/runner.git